+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # changelog
 
-stack-analyze all version and notable changes fixed and new additions in code.
+stack-analyze all version and notable changes, fixed, remove and new additions in code.
+
+## version 1.0.7
+### Added
+- add the new modules:
+    - figlet the main module
+    - jest dev module and testing functions
+- add the options page speed and about the main options
+- add return in a option select except about and exit
+### Fixed
+- the website in blank or website without http:// or https:// return the main options
+- eslint custom rules without style guide
+### Changed
+- changed the inquirer list to inquirer rawlist
+- rewirte jsdoc in all project except test files
+- the tech stack using other console methods console.log only in exit cli
 
 ## version 1.0.6
 ### Added
@@ -11,16 +26,14 @@ stack-analyze all version and notable changes fixed and new additions in code.
 - the list options with inquirer
 - welcome message modified
 
-
 ## version 1.0.5
 ### Changed
-- remove the commander module
 - modify the code and module structure
+### remove
+- remove the commander module
 
 ## version 1.0.4
-
-- fisrt official version in npm
-with the modules:
+- fisrt official version in npm with the modules:
     - wappalyzer
     - inquirer
     - commander